Refactoring: split tests into multiple sources
Separate tests for individual commands ('new', 'take') into their own
files. Common utilities are put to a separate module tests/utils/mod.rs.
